PY: quite a bit of refactoring.
