+++ dolang/codegen.py
@@ -1,7 +1,7 @@
 """
 Extension to ast that allow ast -> python code generation.
 
-:copyright: Copyright 2008 by Armin Ronacher.
+:copyright: Copyright 2008 by Armin Ronacher. Mofified by Pablo Winant.
 :license: BSD.
 """
 from ast import *
